Refactor directories to more flat structure
