--- src/commands/current_weather/mod.rs
@@ -1,4 +1,0 @@
-mod coordinates;
-pub mod get_current_weather;
-mod print_result;
-mod weather;
