--- src/commands/current_weather/print_result.rs
@@ -1,7 +1,0 @@
-use crate::commands::current_weather::weather::Forecast;
-
-pub fn print_results(weather_data: &Forecast) {
-    println!("Temperature: {}", weather_data.main.temp);
-    println!("Feels like:  {}", weather_data.main.feels_like);
-    println!("Wind speed:  {}", weather_data.wind.speed);
-}
