--- src/commands/set_config/mod.rs
@@ -1,1 +1,0 @@
-pub mod update_config_value;
