+++ src/entities/mod.rs
@@ -1,0 +1,2 @@
+pub mod forecast;
+pub mod place;
